requirements.txt
attrs==25.4.0
beautifulsoup4==4.12.3
brotli==1.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cssselect==1.3.0
Deprecated==1.3.1
exceptiongroup==1.3.1
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.3
h11==0.16.0
html2text==2025.4.15
idna==3.11
jieba3k==0.35.1
joblib==1.5.3
lxml==6.0.2
lxml_html_clean==0.4.3
mss==10.1.0
newspaper4k==0.9.4.1
nltk==3.9.2
nodriver==0.38.post1
numpy==2.2.6
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
praw==7.8.1
prawcore==2.4.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyVirtualDisplay==3.0
PyYAML==6.0.3
regex==2026.1.15
requests==2.31.0
requests-file==3.0.1
requests-oauthlib==2.0.0
schedule==1.2.2
Scweet==3.2
selenium==4.39.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
snscrape==0.7.0.20230622
sortedcontainers==2.4.0
soupsieve==2.8.1
textblob==0.19.0
tinysegmenter==0.3
tldextract==5.3.1
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
tweepy==4.16.0
typing_extensions==4.15.0
tzdata==2025.3
update-checker==0.18.0
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==13.1
wrapt==2.0.1
wsproto==1.3.2
